28-02-2019 2:46am
Added the method for htaccess rules. So this method will create a new file if it's not existing else it will just add the rules.

Though one issue is that the newly created .htaccess file ends up being read only. Not sure if it's just an issue with Mac OS X ONLY